requirements.txt
fastapi
uvicorn
streamlit
pydantic
requests
beautifulsoup4
python-dotenv
langchain
langchain-anthropic
langgraph
ollama
tenacity
aiolimiter
gtts
elevenlabs
